refactor team analytics code to reduce boilerplate
