+++ src/container/SessionViewContainer/AnalyticsViewContainer.tsx
@@ -2,8 +2,26 @@ import React from 'react';
 import { Col, Panel } from 'rsuite';
 import IndividualAnalytics from '../../component/IndividualAnalytics';
 import TeamAnalytics from '../../component/TeamAnalytics';
+import { getAccuracyData, getDelayData, getEmgData } from '../../utils/analytic';
 
-const AnalyticsViewContainer: React.FunctionComponent<any> = ({ predictedData, rawData, dancerData }) => {
+interface Props {
+  predictedData: any[];
+  rawData: any[];
+  dancerData: any[];
+}
+
+const AnalyticsViewContainer: React.FunctionComponent<Props> = ({ predictedData, rawData, dancerData }) => {
+  const expectedDeviceData = dancerData.map((data) => ({
+    ...data,
+    device_id: data['device']['id'],
+  }));
+
+  const emgData = getEmgData(expectedDeviceData, rawData, true);
+
+  const delayData = getDelayData(expectedDeviceData, predictedData, true);
+
+  const accuracyData = getAccuracyData(expectedDeviceData, predictedData, true);
+
   const renderIndividualAnalytics = () => {
     return dancerData.map((dancer: any, index: number) => (
       <Col md={24 / dancerData.length} sm={24} key={index}>
@@ -25,7 +43,7 @@ const AnalyticsViewContainer: React.FunctionComponent<any> = ({ predictedData, r
         {renderIndividualAnalytics()}
       </Panel>
       <Panel header={<h4>Team Analytics</h4>} collapsible defaultExpanded>
-        <TeamAnalytics session rawData={rawData} predictedData={predictedData} expectedData={dancerData} />
+        <TeamAnalytics accuracyData={accuracyData} emgData={emgData} delayData={delayData} />
       </Panel>
     </>
   );
